+++ contracts/handlers/utils/HMath.sol
@@ -1,0 +1,32 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity 0.8.10;
+
+import "../HandlerBase.sol";
+
+contract HMath is HandlerBase {
+    function getContractName() public pure override returns (string memory) {
+        return "HMath";
+    }
+
+    function add(uint256 a, uint256 b) external payable returns (uint256 ret) {
+        return a + b;
+    }
+
+    function addMany(
+        uint256[] calldata a
+    ) external payable returns (uint256 ret) {
+        uint256 i;
+        for (; i < a.length; ) {
+            ret += a[i];
+            unchecked {
+                i++;
+            }
+        }
+        return ret;
+    }
+
+    function sub(uint256 a, uint256 b) external payable returns (uint256 ret) {
+        return a - b;
+    }
+}