+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Added
+
+- Support add, addMany, and sub.
+
 ## [2.7.0] - 2023-06-07
 
 ### Added